requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
idna==2.8
python-memcached==1.59
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
